first implementation shift

+++ ui/parameterlistmodel.cpp
@@ -82,7 +82,7 @@ void ParameterListModel::createSystem(QString algName) {
         instantiate(params[0].toInt(), params[1].toInt());
   }else if (signature == "gsf") {
       dynamic_cast<GeneralShapeFormationAlg*>(alg)->
-          instantiate(params[0].toInt(), params[1]);
+          instantiate(params[0].toInt(), params[1].toInt());
   } else if (signature == "ballroomdemo") {
     dynamic_cast<BallroomDemoAlg*>(alg)->
         instantiate(params[0].toInt());
